+++ react-components/src/data-providers/core-dm-provider/cogniteCadNodeSourceWithProperties.ts
@@ -3,15 +3,13 @@
  */
 import { type SourceSelectorV3 } from '@cognite/sdk';
 import { COGNITE_CAD_NODE_SOURCE } from './dataModels';
+import { cogniteDescribableSourceWithProperties } from './cogniteDescribableSourceWithProperties';
 
 export const cogniteCadNodeSourceWithProperties = [
   {
     source: COGNITE_CAD_NODE_SOURCE,
     properties: [
-      'name',
-      'description',
-      'tags',
-      'aliases',
+      ...cogniteDescribableSourceWithProperties[0].properties,
       'object3D',
       'model3D',
       'cadNodeReference',
