+++ react-components/src/data-providers/core-dm-provider/cogniteCadRevisionSourceWithProperties.ts
@@ -4,7 +4,7 @@
 import { type SourceSelectorV3 } from '@cognite/sdk';
 import { COGNITE_CAD_REVISION_SOURCE } from './dataModels';
 
-export const cogniteCadNodeSourceWithPRoperties = [
+export const cogniteCadRevisionSourceWithProperties = [
   {
     source: COGNITE_CAD_REVISION_SOURCE,
     properties: ['status', 'published', 'type', 'model3D', 'revisionId']
