+++ react-components/src/data-providers/core-dm-provider/cogniteObject3dSourceWithProperties.ts
@@ -3,15 +3,13 @@
  */
 import { type SourceSelectorV3 } from '@cognite/sdk';
 import { COGNITE_3D_OBJECT_SOURCE } from './dataModels';
+import { cogniteDescribableSourceWithProperties } from './cogniteDescribableSourceWithProperties';
 
 export const cogniteObject3dSourceWithProperties = [
   {
     source: COGNITE_3D_OBJECT_SOURCE,
     properties: [
-      'name',
-      'description',
-      'tags',
-      'aliases',
+      ...cogniteDescribableSourceWithProperties[0].properties,
       'xMin',
       'xMax',
       'yMin',
