+++ react-components/src/data-providers/core-dm-provider/getCadModelsForInstance.ts
@@ -1,18 +1,16 @@
 /*!
  * Copyright 2024 Cognite AS
  */
-import { type QueryRequest } from '@cognite/sdk';
 import { type TaggedAddResourceOptions } from '../../components/Reveal3DResources/types';
 import { type DmsUniqueIdentifier, type FdmSDK } from '../FdmSDK';
 import {
   COGNITE_CAD_NODE_SOURCE,
   COGNITE_CAD_NODE_VIEW_VERSION_KEY,
-  COGNITE_VISUALIZABLE_SOURCE,
   type CogniteCADNodeProperties,
   CORE_DM_SPACE
 } from './dataModels';
-import { cogniteCadNodeSourceWithProperties } from './cogniteCadNodeSourceWithProperties';
 import { getModelIdFromExternalId, getRevisionIdFromExternalId } from './getCdfIdFromExternalId';
+import { cadModelsForInstanceQuery } from './cadModelsForInstanceQuery';
 
 export async function getCadModelsForInstance(
   instance: DmsUniqueIdentifier,
@@ -41,47 +39,3 @@ export async function getCadModelsForInstance(
     }));
   });
 }
-
-const cadModelsForInstanceQuery = {
-  with: {
-    asset: {
-      nodes: {
-        filter: {
-          and: [
-            {
-              equals: {
-                property: ['node', 'externalId'],
-                value: { parameter: 'instanceExternalId' }
-              }
-            },
-            {
-              equals: {
-                property: ['node', 'space'],
-                value: { parameter: 'instanceSpace' }
-              }
-            }
-          ]
-        }
-      }
-    },
-    object_3ds: {
-      nodes: {
-        from: 'asset',
-        through: {
-          view: COGNITE_VISUALIZABLE_SOURCE,
-          identifier: 'object3D'
-        },
-        direction: 'outwards'
-      }
-    },
-    cad_nodes: {
-      nodes: {
-        from: 'object_3ds',
-        through: { view: COGNITE_CAD_NODE_SOURCE, identifier: 'object3D' }
-      }
-    }
-  },
-  select: {
-    cad_nodes: { sources: cogniteCadNodeSourceWithProperties }
-  }
-} as const satisfies Omit<QueryRequest, 'parameters' | 'cursor'>;
